requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
django-otp==1.1.6
django-robots==5.0
django-taggit==3.1.0
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
